Rename package folder `example` to `examples-react` to match package name
