+++ packages/examples-react/package.json
