+++ packages/examples-react/public/favicon.ico
