+++ packages/examples-react/public/manifest.json
